MS Teams JSON
