+++ lib/puptime/notifier/teams.json
@@ -1,0 +1,6 @@
+{
+  "teams": {
+    "title": "Puptime Notification",
+    "text": "message"
+  }
+}
